Extract bob80333's stick smoothing algo into dedicated class.
This is an excellent example of extracting code for the purpose of
reuse. The smoothing algorithm can now be used for any stick readings.

The next step will be to use this for the local code within
`TeleopDrive`. The team should also decide if it makes sense to use it
for elevator and arm input.