Use `DeferredCmdSupplier` to resolve chicken/egg problem
The default Commands depend on the related Subsystem and the Subsystems
depend on their default Commands.  This circular dependency crates a
chicken/egg problem (e.g. Eggs come from chickens, but chickens come
from eggs!)

`DeferredCmdSupplier` gives us a way to solve this problem. See the
comments inside the `DeferredCmdSupplier` class for details of the
problem and how it is solved.